Lessen_dissolve_item_trap_potential
Changing it from 5% chance per item to 3% per item, and also adding in a 20% chance to break out of the for loop when checking inventory.
None of this even triggers if the player saves vs traps, so it's rare. But I think it needs to be less harsh than it currently is anyway, even on a failed save.