Raise_level_req_for_Resurrect-Bloodfusiion
Spell was originally designed for a 30 level system. Res/Bloodf slaves was an unforeseen consequence. Going back i wouldn't even have put this in. lol..No other changes to it, but i'd like them to level to 40 if they going to res/bloodf slave. Very powerful spell. I set it to MAXALVL if we ever increase from 40 going forward.