Update login.cpp (#62)
Request by Nik to add divine-weapon and arcane-weapon. This will be used for stuff like a "rod of blasting" or a "rod of holy might". He's already made several new items and has had to put them to subtype bow.

These subtypes will be class restricted for choice, and only will be allowed to be picked in weapon training if level 4+. so they cant be grabbed at char creation.