Update_armor_spell_and_starting_spell_changes
Updated armor spell. It was crapping out after one hit the way it was before.
It will now work as intended. Details are commented in the code. Players can see the remaining strength of their armor spell in the "effects" command.

Mages (as their first class) and liches now start with the armor and magic-missile spells.

Also set up clerics and druids to always start with the vigor spell.

Also added starting weapons for the recently added knife and sling skills.