Remove clear long and clear short
Back during the dial-up days, clearing long and short descriptions improved throughput and reduced lag. This is no longer necessary or needed with modern day connections.

Staff unanimously agrees it is time for this option to go the way of the dodo bird.

Flags have been zeroed out and reclaimed.  Players will update when they log in. The version has been bumped to 2.5.2.