Hold spell refactor, racial quirks first pass, many other additions/fixes (#193)
* racial quirks and other stuff  (work in progress)

-hold-person spell updates (work in progress)
-racial quirks in combat system
-some other random stuff

* Update cmd.cpp

-Fix output of "help commands"
-Added descriptions to all relevant legacy commands that were missing and altered/added some others.

* -added isMagicallyHeld() checks everywhere

* Test commit with stupid MS line problem

* -further worked on hold spells, got them working basically. Have to polish next.

* -Rounded out hold spells, added more of them, added some new functionality in code around monTypes. All left to do is tweak the durations of the hold effect and tweak. Ready for minor playtesting. Likely get Nik to help with that.

* -Rounded out hold spells, added more of them, added some new functionality in code around monTypes. All left to do is tweak the durations of the hold effect and tweak. Ready for minor playtesting. Likely get Nik to help with that.

- Output fixes
- Added greater-invis for *gcast
- disabled hold-person spell except for play testers

* -Debugged mobs casting hold spells
-Added mtypes and sexes to bard/mage conjures
-Tweaked hold spells further and debugged outputs and other found flow problems
-Fixed M_WILL_KILL_MASTER mflag for mobs attacking pets

* -General cleanup

* - More cleanup
- Added checkResistEnchantments() to scare/fear

* -Disabled mobs casting hold spells except for vs p-testers and other mobs
-Hold spells only available to p-testers for now
-Added gore attack for minotaurs and existing minotaurs
-Reduced chances of weapons and armor losing shots
-Kick timer reduced by 3 seconds
-Adjusted List to separate .area and .id into different columns
-Fixed ambush bug. It uses shots the same as backstab now

* -Fixed output error for when mobs attack you for healing their target

* -Fixed output sequencing error with pet attack. Now pet will stop attacking only if it has an enemy, whereas before, always stopped attacking. If no enemy, will ask player what target the pet should attack.

* -Fixed output sequencing error with pet attack. Now pet will stop attacking only if it has an enemy, whereas before, always stopped attacking. If no enemy, will ask player what target the pet should attack.

* -Refactored Creature::checkResistEnchantments() and turned it into multiple functions for better digestibility.
Had to do extensive re-testing for outputs, etc..and this helped because also found a few bugs I fixed. Hold spells will remain in p-testing into next year probably.
Even with the refactor, it's still a little busy, but IMO due to the nature of all the checks the hold spells have to do, any more condensing is just going to make things just as complex as they already are with really no gain that I can see. It definitely is easier to follow now. You may also notice a couple redundant calls, such as checking if an enchantment spell is resistable in all the broken out functions. Reason behind that is I wanted to make the functions able to be called individually if necessary for whatever future need might come up for that.
- Also as suggested, I removed all of the if(isEffected("blah") removeEffect("blah") that *I* coded in, since removeEffect("blah") checks for existing "blah" effect already.
- I also added 3 new monTypes: OOZE, MODRON, and DAEMON. Also changed some minor output in displayCreature() A couple minor additions - not very significant changes.

All the other suggestions you provided were also already completed where necessary in commits before this one.

* -Refactored Creature::checkResistEnchantments() and turned it into multiple functions for better digestibility.
Had to do extensive re-testing for outputs, etc..and this helped because also found a few bugs I fixed. Hold spells will remain in p-testing into next year probably.
Even with the refactor, it's still a little busy, but IMO due to the nature of all the checks the hold spells have to do, any more condensing is just going to make things just as complex as they already are with really no gain that I can see. It definitely is easier to follow now. You may also notice a couple redundant calls, such as checking if an enchantment spell is resistable in all the broken out functions. Reason behind that is I wanted to make the functions able to be called individually if necessary for whatever future need might come up for that.
- Also as suggested, I removed all of the if(isEffected("blah") removeEffect("blah") that *I* coded in, since removeEffect("blah") checks for existing "blah" effect already.
- I also added 3 new monTypes: OOZE, MODRON, and DAEMON. Also changed some minor output in displayCreature() A couple minor additions - not very significant changes.

All the other suggestions you provided were also already completed where necessary in commits before this one.

* -Refactored Creature::checkResistEnchantments() and turned it into multiple functions for better digestibility.
Had to do extensive re-testing for outputs, etc..and this helped because also found a few bugs I fixed. Hold spells will remain in p-testing into next year probably.
Even with the refactor, it's still a little busy, but IMO due to the nature of all the checks the hold spells have to do, any more condensing is just going to make things just as complex as they already are with really no gain that I can see. It definitely is easier to follow now. You may also notice a couple redundant calls, such as checking if an enchantment spell is resistable in all the broken out functions. Reason behind that is I wanted to make the functions able to be called individually if necessary for whatever future need might come up for that.
- Also as suggested, I removed all of the if(isEffected("blah") removeEffect("blah") that *I* coded in, since removeEffect("blah") checks for existing "blah" effect already.
- I also added 3 new monTypes: OOZE, MODRON, and DAEMON. Also changed some minor output in displayCreature() A couple minor additions - not very significant changes.

All the other suggestions you provided were also already completed where necessary in commits before this one.

-removed some superfluous comments i forgot about