Gold luck fix (#198)
* -Refactored Creature::checkResistEnchantments() and turned it into multiple functions for better digestibility.
Had to do extensive re-testing for outputs, etc..and this helped because also found a few bugs I fixed. Hold spells will remain in p-testing into next year probably.
Even with the refactor, it's still a little busy, but IMO due to the nature of all the checks the hold spells have to do, any more condensing is just going to make things just as complex as they already are with really no gain that I can see. It definitely is easier to follow now. You may also notice a couple redundant calls, such as checking if an enchantment spell is resistable in all the broken out functions. Reason behind that is I wanted to make the functions able to be called individually if necessary for whatever future need might come up for that.
- Also as suggested, I removed all of the if(isEffected("blah") removeEffect("blah") that *I* coded in, since removeEffect("blah") checks for existing "blah" effect already.
- I also added 3 new monTypes: OOZE, MODRON, and DAEMON. Also changed some minor output in displayCreature() A couple minor additions - not very significant changes.

All the other suggestions you provided were also already completed where necessary in commits before this one.

-removed some superfluous comments i forgot about

* -Remove gold check in player->computeLuck()
-bump version to 2.58b